moving sKCSD stuff around
