+++ data/morphology/Badea2011Fig2Du.CNG.swc
