+++ data/morphology/ElcoordsDomi14.txt
