+++ data/morphology/Mainen_swcLike.swc
