+++ data/morphology/ballstick.hoc
