+++ data/morphology/ballstick.swc
