+++ data/sinsyn.mod
