+++ figures/sKCSD_paper/Data/ball_and_stick_8/electrode_positions/elcoord_x_y_z
