+++ figures/sKCSD_paper/Data/ball_and_stick_8/morphology/ball_and_stick.swc
