+++ figures/sKCSD_paper/Data/gang_7x7_200/electrode_positions/elcoord_x_y_z
