+++ figures/sKCSD_paper/Data/gang_7x7_200/morphology/Badea2011Fig2Du.CNG.swc
