+++ kcsd/__init__.py
@@ -1,5 +1,5 @@
 from .KCSD import KCSD1D, KCSD2D, KCSD3D, MoIKCSD
-from . import sKCSD, sKCSDcell
+from . import sKCSD
 from .validation import csd_profile
 from .validation.ValidateKCSD import ValidateKCSD, ValidateKCSD1D, ValidateKCSD2D, ValidateKCSD3D, SpectralStructure
 from .validation.VisibilityMap import VisibilityMap1D, VisibilityMap2D, VisibilityMap3D
